requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
ln-markets==2.0.10
numpy==2.3.2
pandas==2.3.1
pyaml==25.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
tradingview-ta==3.3.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
